requirements.txt
Django>=1.8
djangorestframework
pymongo
django-cors-headers
werkzeug
PyJWT
requests